+++ buildVars.py
@@ -2,7 +2,7 @@ addon_info = {
 	"addon_name": "folderTextFinder",
 	"addon_summary": "Folder Text Finder",
 	"addon_description": "Search files containing text in the current File Explorer folder.",
-	"addon_version": "0.2.1",
+	"addon_version": "0.2.2",
 	"addon_author": "Tara",
 	"addon_url": "https://github.com/oasisfan277/folder-text-finder",
 	"addon_docFileName": "readme.html",